Create README.md
Trended Renewal Processes are a suitable model for studying the reliability of repairable systems under the assumption of a general repair level. If the renewal distribution in the transformed time domain corresponds to a Gamma distribution, the process is determined with three nonnegative parameters. In this code, failure times for a repairable system with a general repair level are simulated, and the corresponding parameters are estimated.